translearn-pptmaker/requirements.txt
aiofiles==0.5.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
contextvars==2.4
dataclasses==0.7
fastapi==0.60.1
googletrans==3.0.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.29
httpcore==0.9.1
httptools==0.1.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.14
lxml==4.5.2
Pillow==7.2.0
pydantic==1.6.1
python-pptx==0.6.18
requests==2.24.0
rfc3986==1.4.0
sniffio==1.1.0
starlette==0.13.6
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
XlsxWriter==1.3.0
